+++ src/Stage.js
@@ -413,6 +413,10 @@
       this._fire(CONTENT_MOUSEOUT, { evt: evt });
     },
     _mousemove: function(evt) {
+      // workaround for mobile IE to force touch event when unhandled pointer event elevates into a mouse event
+      if (Konva.UA.ieMobile) {
+        return this._touchmove(evt);
+      }
       this._setPointerPosition(evt);
       var shape;
 
@@ -461,6 +465,10 @@
       }
     },
     _mousedown: function(evt) {
+      // workaround for mobile IE to force touch event when unhandled pointer event elevates into a mouse event
+      if (Konva.UA.ieMobile) {
+        return this._touchstart(evt);
+      }
       this._setPointerPosition(evt);
       var shape = this.getIntersection(this.getPointerPosition());
 
@@ -490,6 +498,10 @@
       // }
     },
     _mouseup: function(evt) {
+      // workaround for mobile IE to force touch event when unhandled pointer event elevates into a mouse event
+      if (Konva.UA.ieMobile) {
+        return this._touchend(evt);
+      }
       this._setPointerPosition(evt);
       var shape = this.getIntersection(this.getPointerPosition()),
         clickStartShape = this.clickStartShape,